refactor(reward-memory): separate projection validation
